split up field context code
