+++ src/core/query/field_context/EmbeddedFieldContext.ts
@@ -1,0 +1,15 @@
+import {FieldContext} from "./FieldContext"
+
+export class EmbeddedFieldContext extends FieldContext {
+
+  getAggregationPath(){
+    return undefined
+  }
+
+  wrapAggregations(...aggregations){
+    return aggregations
+  }
+  wrapFilter(filter){
+    return filter
+  }
+}
