+++ src/core/query/field_context/FieldOptions.ts
@@ -1,0 +1,5 @@
+export interface FieldOptions {
+  type:String,
+  field?:String,
+  options?:any
+}
