+++ src/core/query/index.ts
@@ -1,4 +1,4 @@
 export * from "./ImmutableQuery"
 export * from "./query_dsl"
 export * from "./SelectedFilter"
-export * from "./field_context/FieldContext"
+export * from "./field_context"
