+++ tsconfig.json
@@ -203,7 +203,13 @@
     "./src/core/accessors/ViewOptionsAccessor.ts",
     "./src/core/history.ts",
     "./src/core/index.ts",
+    "./src/core/query/field_context/ChildrenFieldContext.ts",
+    "./src/core/query/field_context/EmbeddedFieldContext.ts",
     "./src/core/query/field_context/FieldContext.ts",
+    "./src/core/query/field_context/FieldContextFactory.ts",
+    "./src/core/query/field_context/FieldOptions.ts",
+    "./src/core/query/field_context/index.ts",
+    "./src/core/query/field_context/NestedFieldContext.ts",
     "./src/core/query/ImmutableQuery.ts",
     "./src/core/query/index.ts",
     "./src/core/query/query_dsl/aggregations/AggsContainer.ts",
